Language Definition Update
In order to improve the definition of validation rules (and other
activities related to working with the language implementation-wise),
the individual rules for declaring artifacts, fragments, and results are
replaced by a single rule for element declarations, which in turn relies
on a rule that defines the generic type (one of artifact, fragment, or
result).
Further, the entry rule now differentiates between element declarations
and change identifier assignments. This eases the access to these two
different concepts in the language as an analysis definition now
provides individual getter for them.