External Language Elements Integration Update
Addition of the language registry and the language creator as well as
the required annotations for artifact, fragment, and result parameter
types.
Further modifications to the language element provider to connect the
new classes, in particular, extraction of the "ExternalElementException"
from the provider to a dedicated class as we need it also in the
language creator.