Integration Tests
Adaptation of the basic language registry tests, which now require
method objects for the creation of call instances. This adaptation may
cause failures in future, if we check for methods that are annotated
with multiple annotations, such that the represent extractor and
analysis calls at the same time, which is not possible.