External Language Elements Integration Update
Update of the checks for duplicates, if a new element is added to the
language registry. These checks now use the new equals-methods of the
language elements and consider all hash maps containing equal language
elements, e.g., to avoid the same method to introduce an extractor and
an analysis call.