External Language Elements Integration Update
First part of the fix for the detection of language elements. We checked
at the wrong point for duplicates when some lists were already removed
from hash maps. Hence, no duplicates in that lists could be found.
Simply moving the duplicate-check before removing lists from the maps
fixes this issue.
However, some unit tests still fail due to positive equality-checks,
which are in general correct, but not in the scope of adding language
elements. Fix will follow.