Integration Tests
Update of the unit tests for checking correct rejection of duplicated
language elements when adding them to the language registry. We are now
not that strict (e.g., we allow the same parameter type as being an
artifact and a fragment type). However, we check that the same method
cannot declare, e.g., an extractor and an analysis call at the same
time.