External Language Elements Integration Update
Addition of further validations, in particular, for the availability of
return types and parameters of calls and assignable elements of change
identifiers. However, some tests fail due to these new validations,
which reject the addition of some calls or change identifiers with
attributes that are not available during the tests. Fixes will follow
next.