Integration Tests
Fixing some test that fail due to the new validation methods in the
registry. The reason for failing is simply that the defined elements for
testing use parameter types, which are not available in the registry at
the point the tests are executed. With a change of these types to those
available in the registry, the tests run successfully again.