Validation Rules
Addition of change identifier assignment validations, corresponding
methods in the language registry to check for availability of certain
change identifiers, and a method in the change identifier class to check
for acceptance of elements to which a change identifier can be assigned
to.
However, two parsing tests fail as the used models are no longer passing
the validation. This is because of using unavailable change identifiers.
Will be fixes soon.