Validation Rules
Addition and integration of new utility functions for resolving
operations, calls, etc. to their actual (return) type, which is used to
compare such types with the other elements to which they are assigned.

Tests still fail, which has to be fixed by adapting the test operations
and add them as language elements to the language registry before test
execution.