External Language Elements Integration Update
Revert the previous commit and add the extra properties for member
operation as a special case in the Call class. This eases the creation
of such elements and their management as most of the already available
management capabilities rely on the element type, which would require
extra effort and functionalities to handle the previously new
MemberOperation class.