External Language Elements Integration Update
Detection of change identifier duplicates must be performed in a way
that is independent of whether a list of change identifiers with the
current name already exists. If not, no duplication check will be
performed, if there is no list missing that there might be a duplicate
in the cached change identifier list (that check was not performed
before, if the was no list for the current change identifier name).