LEA Loading
Revert last commit and leave loading of language elements to external
component (in the current case the ui). For that purpose, restrict calls
to the language element provider, which no offers a method for loading
core language elements only. The core language element provider is not
visible outside the package anymore.
Make the utility package accessible for other plug-ins, e.g., to use the
logger.